refactor(json): move json module to top-level
