+++ json/common.ts
@@ -1,0 +1,17 @@
+// Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
+/** The type of the result of parsing JSON. */
+export type JsonValue =
+  | { [key: string]: JsonValue | undefined }
+  | JsonValue[]
+  | string
+  | number
+  | boolean
+  | null;
+
+/** Optional object interface for `JSONParseStream` and `ConcatenatedJsonParseStream`. */
+export interface ParseStreamOptions {
+  /** Controls the buffer of the TransformStream used internally. Check https://developer.mozilla.org/en-US/docs/Web/API/TransformStream/TransformStream. */
+  readonly writableStrategy?: QueuingStrategy<string>;
+  /** Controls the buffer of the TransformStream used internally. Check https://developer.mozilla.org/en-US/docs/Web/API/TransformStream/TransformStream. */
+  readonly readableStrategy?: QueuingStrategy<JsonValue>;
+}
